fix/convert to float32 from bfloat16

+++ agentenv/examples/behavioral_cloning/train_behavioral_clone.py
@@ -86,7 +86,7 @@ def main():
 
     tokenizer = AutoTokenizer.from_pretrained(args.model_train_path)
     model = AutoModelForCausalLM.from_pretrained(
-        args.model_train_path, low_cpu_mem_usage=True, torch_dtype=torch.bfloat16
+        args.model_train_path, low_cpu_mem_usage=True, torch_dtype=torch.float32
     )
     model.gradient_checkpointing_enable()
